# [performance] Add live pages