--- a/pages/performance/table-mui.js
+++ b/pages/performance/table-mui.js
@@ -1,0 +1,45 @@
+/* eslint-disable react/no-array-index-key */
+
+import React from 'react';
+import NoSsr from '@material-ui/core/NoSsr';
+import Table from '@material-ui/core/Table';
+import TableBody from '@material-ui/core/TableBody';
+import TableCell from '@material-ui/core/TableCell';
+import TableHead from '@material-ui/core/TableHead';
+import TableRow from '@material-ui/core/TableRow';
+
+const data = { name: 'Frozen yoghurt', calories: 159, fat: 6.0, carbs: 24, protein: 4.0 };
+const rows = Array.from(new Array(100)).map(() => data);
+
+function TableMui() {
+  return (
+    <NoSsr defer>
+      <Table>
+        <TableHead>
+          <TableRow>
+            <TableCell>Dessert (100g serving)</TableCell>
+            <TableCell>Calories</TableCell>
+            <TableCell>Fat (g)</TableCell>
+            <TableCell>Carbs (g)</TableCell>
+            <TableCell>Protein (g)</TableCell>
+          </TableRow>
+        </TableHead>
+        <TableBody>
+          {rows.map((row, index) => (
+            <TableRow key={index}>
+              <TableCell component="th" scope="row">
+                {row.name}
+              </TableCell>
+              <TableCell>{row.calories}</TableCell>
+              <TableCell>{row.fat}</TableCell>
+              <TableCell>{row.carbs}</TableCell>
+              <TableCell>{row.protein}</TableCell>
+            </TableRow>
+          ))}
+        </TableBody>
+      </Table>
+    </NoSsr>
+  );
+}
+
+export default TableMui;