I have added new stress test pages so we can more easily benchmark the changes. I would like to create a way to automate the tests, sample many iterations. All we have with those pages, is an order of magnitude as well as a debug env. It's not accurate enough!
- raw: 15ms
- component: 25ms
- hook: 65ms
- sc: 70ms
- emotions: 80ms
- mui: 120ms